+++ django_elastic_migrations/models.py
@@ -30,7 +30,7 @@ class Index(models.Model):
     """
     name = models.CharField(verbose_name="Index Name", max_length=32, unique=True)
 
-    # Django convention is to use '+' for related name when you don't need the
+    # Django's convention is to use '+' for related name when you don't need the
     # reverse relation. in this case, we already have IndexVersion pointing
     # back to Index, so we don't need that reverse name.
     # See https://docs.djangoproject.com/en/2.0/ref/models/fields/#django.db.models.ForeignKey.related_name